server/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
mediapipe==0.10.5
requests==2.31.0

# Added for Google Gemini visual analysis
google-generativeai>=0.3.0
python-dotenv>=0.20.0

# Image Quality Assessment
piq==0.7.0
torch>=1.7.1
torchvision>=0.8.2
open-clip-torch==2.32.0

